fix radium warnings
