+++ src/components/sightsListItem/index.jsx
@@ -36,7 +36,8 @@ const styles = {
   },
   reset: {
     padding: 0,
-    margin: 0,
+    marginTop: 0,
+    marginBottom: 0,
   },
   title: {
     fontSize: "16px",
